Fix: harden g-6 tests with deterministic ids, storage state, and API split
