+++ tests/api/platform/g-6-volunteer-contract-boundary-and-regression-hardening.atdd.api.spec.ts
@@ -1,159 +1,13 @@
 import { apiRequest } from '../../support/helpers/apiClient';
 import { test, expect } from '../../support/fixtures/connectShyftStoryG6.fixture';
-
-type QueueItem = {
-  threadId?: string;
-  bucket?: string;
-  state?: string;
-  voicemailIndicator?: boolean;
-  display?: {
-    title?: string;
-    preview?: string;
-    timestampLabel?: string;
-    urgencyLabel?: string;
-    stateLabel?: string;
-    contextPills?: string[];
-    voicemailLabel?: string;
-    inboundContext?: string;
-    outboundContext?: string;
-    neighborContext?: string;
-    conferenceContext?: string;
-    threadId?: string;
-    priorityRank?: number;
-    rawStateChip?: string;
-    routingMetadata?: unknown;
-    webhookMetadata?: unknown;
-    systemMetadata?: unknown;
-  };
-};
-
-type QueueEnvelope = {
-  ok?: boolean;
-  code?: string;
-  data?: {
-    items?: QueueItem[];
-  };
-};
-
-type ThreadDetailEnvelope = {
-  ok?: boolean;
-  code?: string;
-  errorType?: string;
-  data?: {
-    thread?: {
-      threadId?: string;
-      state?: string;
-      timeline?: Array<{
-        conversationType?: string;
-        renderMode?: string;
-        firstClass?: boolean;
-      }>;
-    };
-    lifecycle?: {
-      priorState?: string;
-      nextState?: string;
-      reopenedByInbound?: boolean;
-      reopenedFromClosed?: boolean;
-      sameThreadId?: boolean;
-      noInboundAutoReopenSideEffects?: boolean;
-    };
-    uiFeedback?: {
-      severity?: string;
-      ariaLive?: string;
-      presentation?: string;
-      message?: string;
-    };
-    feedback?: {
-      taxonomy?: string;
-    };
-  };
-  refusalType?: string;
-};
-
-type InboundWebhookEnvelope = {
-  ok?: boolean;
-  code?: string;
-  data?: {
-    thread?: {
-      threadId?: string;
-      state?: string;
-    };
-    lifecycle?: {
-      reopenedByInbound?: boolean;
-    };
-    timeline?: {
-      routingDecision?: string;
-    };
-  };
-};
-
-const UUID_PATTERN =
-  /[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}/i;
-
-const collectPrimaryCopy = (item: QueueItem): string[] => {
-  const fields: string[] = [];
-
-  const collect = (value: unknown): void => {
-    if (typeof value !== 'string') {
-      return;
-    }
-
-    const normalized = value.trim();
-    if (normalized.length > 0) {
-      fields.push(normalized);
-    }
-  };
-
-  collect(item.display?.title);
-  collect(item.display?.preview);
-  collect(item.display?.timestampLabel);
-  collect(item.display?.urgencyLabel);
-  collect(item.display?.stateLabel);
-  collect(item.display?.voicemailLabel);
-  collect(item.display?.inboundContext);
-  collect(item.display?.outboundContext);
-  collect(item.display?.neighborContext);
-  collect(item.display?.conferenceContext);
-
-  if (Array.isArray(item.display?.contextPills)) {
-    for (const pill of item.display.contextPills) {
-      collect(pill);
-    }
-  }
-
-  return fields;
-};
-
-const readItems = (body: QueueEnvelope): QueueItem[] => {
-  if (!Array.isArray(body.data?.items)) {
-    return [];
-  }
-
-  return body.data.items;
-};
-
-const resolveFeedbackTaxonomy = (
-  body: ThreadDetailEnvelope,
-): 'success' | 'refusal' | 'error' | null => {
-  const explicitTaxonomy = body.data?.feedback?.taxonomy ?? body.data?.uiFeedback?.severity;
-  if (explicitTaxonomy === 'success' || explicitTaxonomy === 'refusal' || explicitTaxonomy === 'error') {
-    return explicitTaxonomy;
-  }
-
-  if (body.ok === true) {
-    return 'success';
-  }
-
-  if (body.refusalType === 'business') {
-    return 'refusal';
-  }
-
-  if (body.ok === false || body.errorType === 'system') {
-    return 'error';
-  }
-
-  return null;
-};
+import {
+  collectPrimaryCopy,
+  type QueueEnvelope,
+  readItems,
+  type ThreadDetailEnvelope,
+  UUID_PATTERN,
+} from './g-6-volunteer-contract-boundary-and-regression-hardening.shared';
+import './g-6-volunteer-contract-boundary-and-regression-hardening.atdd.api.lifecycle-and-feedback.cases';
 
 test.describe('Story g.6 Volunteer Contract Boundary and Regression Hardening (ATDD API RED)', () => {
   test(
@@ -300,108 +154,6 @@ test.describe('Story g.6 Volunteer Contract Boundary and Regression Hardening (A
         ariaLive: 'polite',
         presentation: 'contextual-action-feedback',
       });
-    },
-  );
-
-  test(
-    '[G6-ATDD-API-004][P0] inbound webhook activity on CLOSED threads keeps state locked and routes fallback without auto-reopen @P0',
-    async ({
-      request,
-      storyG6Context,
-      storyG6AdminHeaders,
-      storyG6InboundClosedPayload,
-    }) => {
-      const webhookResponse = await apiRequest(request, {
-        method: 'POST',
-        path: storyG6Context.paths.inboundWebhook,
-        headers: storyG6AdminHeaders,
-        data: storyG6InboundClosedPayload,
-      });
-
-      expect(webhookResponse.status()).toBe(200);
-      const webhookBody = (await webhookResponse.json()) as InboundWebhookEnvelope;
-      expect(webhookBody).toMatchObject({
-        ok: true,
-        code: 'CONNECTSHYFT_WEBHOOK_ACCEPTED',
-        data: {
-          thread: {
-            threadId: storyG6Context.threadIds.closedInbound,
-            state: 'CLOSED',
-          },
-          lifecycle: {
-            reopenedByInbound: false,
-          },
-          timeline: {
-            routingDecision: 'intake_fallback',
-          },
-        },
-      });
-
-      const detailResponse = await apiRequest(request, {
-        method: 'GET',
-        path: `${storyG6Context.paths.threads}/${storyG6Context.threadIds.closedInbound}`,
-        headers: storyG6AdminHeaders,
-      });
-
-      expect(detailResponse.status()).toBe(200);
-      const detailBody = (await detailResponse.json()) as ThreadDetailEnvelope;
-      expect(detailBody.data?.thread?.state).toBe('CLOSED');
-    },
-  );
-
-  test(
-    '[G6-ATDD-API-005][P1] volunteer action feedback contracts publish explicit success/refusal/error taxonomy without contradictory messaging @P1',
-    async ({
-      request,
-      storyG6Context,
-      storyG6VolunteerHeaders,
-      storyG6OutboundCallPayload,
-    }) => {
-      const successResponse = await apiRequest(request, {
-        method: 'POST',
-        path: `${storyG6Context.paths.threads}/${storyG6Context.threadIds.closedOutbound}/call`,
-        headers: storyG6VolunteerHeaders,
-        data: storyG6OutboundCallPayload,
-      });
-      const refusalResponse = await apiRequest(request, {
-        method: 'POST',
-        path: `${storyG6Context.paths.threads}/${storyG6Context.threadIds.unclaimedPrefersNo}/messages`,
-        headers: storyG6VolunteerHeaders,
-        data: {
-          orgUnitId: storyG6Context.orgUnitId,
-          channel: 'sms',
-          body: 'Outbound attempt without override metadata',
-        },
-      });
-      const errorResponse = await apiRequest(request, {
-        method: 'POST',
-        path: storyG6Context.paths.threads,
-        headers: storyG6VolunteerHeaders,
-        data: {
-          orgUnitId: storyG6Context.orgUnitId,
-          neighborId: storyG6Context.neighborIds.closedInbound,
-          source: 'VOICE',
-          threadId: 'thread-g6-client-forbidden',
-          lastInboundCsNumberId: 'cs-number-ux-r4-403',
-          preferredOutboundCsNumberId: 'cs-number-ux-r4-503',
-        },
-      });
-
-      expect(successResponse.status()).toBe(200);
-      expect(refusalResponse.status()).toBe(200);
-      expect(errorResponse.status()).toBe(400);
-
-      const successBody = (await successResponse.json()) as ThreadDetailEnvelope;
-      const refusalBody = (await refusalResponse.json()) as ThreadDetailEnvelope;
-      const errorBody = (await errorResponse.json()) as ThreadDetailEnvelope;
-
-      expect(resolveFeedbackTaxonomy(successBody)).toBe('success');
-      expect(refusalBody.refusalType).toBe('business');
-      expect(resolveFeedbackTaxonomy(refusalBody)).toBe('refusal');
-      expect(errorBody.refusalType).toBe('client');
-      expect(resolveFeedbackTaxonomy(errorBody)).toBe('error');
-      expect(String(refusalBody.data?.uiFeedback?.message ?? '')).not.toMatch(/reopened|auto-reopen/i);
-      expect(String(errorBody.data?.uiFeedback?.message ?? '')).not.toMatch(/reopened|auto-reopen/i);
     },
   );
 });
